+++ executor/memtlb.h
@@ -41,7 +41,7 @@ private:
     FEATURE_SMALL_PDPT = 1 << 3,
     FEATURE_LONG       = 1 << 4,
   };
-  unsigned (*tlb_fill_func)(MemTlb *tlb, unsigned long virt, unsigned type, long unsigned &phys);
+  unsigned (*tlb_fill_func)(MemTlb *tlb, uintptr_t virt, unsigned type, uintptr_t &phys);
 
 #define AD_ASSIST(bits)							\
   if ((pte & (bits)) != (bits))						\
@@ -54,12 +54,12 @@ private:
     }
 
   template <unsigned features, typename PTE_TYPE>
-    static unsigned tlb_fill(MemTlb *tlb, unsigned long virt, unsigned type, long unsigned &phys)
+    static unsigned tlb_fill(MemTlb *tlb, uintptr_t virt, unsigned type, uintptr_t &phys)
   {  return tlb->tlb_fill2<features, PTE_TYPE>(virt, type, phys); }
 
 
   template <unsigned features, typename PTE_TYPE>
-    unsigned tlb_fill2(unsigned long virt, unsigned type, long unsigned &phys)
+    unsigned tlb_fill2(uintptr_t virt, unsigned type, uintptr_t &phys)
   {
     PTE_TYPE pte;
     if (features & FEATURE_SMALL_PDPT) pte = _pdpt[(virt >> 30) & 3]; else pte = READ(cr3);
@@ -115,7 +115,7 @@ private:
     return _fault;
   }
 
-  int virt_to_phys(unsigned long virt, Type type, long unsigned &phys) {
+  int virt_to_phys(uintptr_t virt, Type type, uintptr_t &phys) {
 
     if (tlb_fill_func) return tlb_fill_func(this, virt, type, phys);
     phys = virt;
@@ -125,9 +125,9 @@ private:
   /**
    * Find a CacheEntry to a virtual memory access.
    */
-  CacheEntry *find_virtual(unsigned virt, unsigned len, Type type) {
+  CacheEntry *find_virtual(uintptr_t virt, size_t len, Type type) {
 
-    unsigned long phys1, phys2;
+    uintptr_t phys1, phys2;
     if (!virt_to_phys(virt, type, phys1)) {
       if (!((virt ^ (virt + len - 1)) & ~0xfff)) phys2 = ~0ul;
       else
@@ -180,7 +180,7 @@ protected:
   /**
    * Read the len instruction-bytes at the given address into a buffer.
    */
-  int read_code(unsigned long virt, unsigned len, void *buffer)
+  int read_code(uintptr_t virt, size_t len, void *buffer)
   {
     assert(len < 16);
     CacheEntry *entry = find_virtual(virt & ~3, (len + (virt & 3) + 3) & ~3ul, user_access(Type(TYPE_X | TYPE_R)));
@@ -195,7 +195,7 @@ protected:
   }
 
 
-  int prepare_virtual(unsigned virt, unsigned len, Type type, void *&ptr)
+  int prepare_virtual(uintptr_t virt, size_t len, Type type, void *&ptr)
   {
     bool round = (virt | len) & 3;
     CacheEntry *entry = find_virtual(virt & ~3ul, (len + (virt & 3) + 3) & ~3ul, round ? Type(type | TYPE_R) : type);
