+++ include/executor/bios.h
@@ -37,7 +37,7 @@ protected:
   /**
    * Write bios data helper.
    */
-  void write_bda(unsigned short offset, unsigned value, unsigned len)
+  void write_bda(unsigned short offset, unsigned value, size_t len)
   {
     assert(len <= sizeof(value));
     copy_out(0x400 + offset, &value, len);
@@ -47,7 +47,7 @@ protected:
   /**
    * Read bios data helper.
    */
-  unsigned read_bda(unsigned offset)
+  unsigned read_bda(size_t offset)
   {
     unsigned res;
     copy_in(0x400 + offset, &res, sizeof(res));
