+++ model/ioapic.cc
@@ -144,7 +144,7 @@ private:
 
 	_rirr[pin] = level;
 	_ds[pin]   = false;
-	unsigned long phys = MessageMem::MSI_ADDRESS | (dst >> 12) & 0xffff0;
+	uintptr_t phys = MessageMem::MSI_ADDRESS | (dst >> 12) & 0xffff0;
 	if (value & MessageApic::ICR_DM) phys |= MessageMem::MSI_DM;
 	if ((value & 0x700) == 0x100)    phys |= MessageMem::MSI_RH;
 	if (_rirr[pin])                 value |= 1 << 14;
@@ -234,7 +234,7 @@ public:
 
   void discovery() {
 
-    unsigned length = discovery_length("APIC", 44);
+    size_t length = discovery_length("APIC", 44);
 
     if (!_gsibase) {
       // override IRQ 0->2
@@ -256,7 +256,7 @@ public:
 
 
 
-  IOApic(Motherboard &mb, unsigned long base, unsigned gsibase) : _mb(mb), _base(base), _gsibase(gsibase)
+  IOApic(Motherboard &mb, uintptr_t base, unsigned gsibase) : _mb(mb), _base(base), _gsibase(gsibase)
   {
     reset();
     _mb.bus_mem.add(this,       receive_static<MessageMem>);
